Dynamically determine vehicle name for ROS
The shipping build creates a unique name for each vehicle that is spawned which
we rely on for controlling the vehicle. This will use the state sensor to query
the vehicle's name and use it to control it after its spawned.

To test, run the ROS example with the shipping build to see that it works.