Added configuration and demo for waypoint debug
This adds the configuration values and an example of how to use the
waypoint sensor's debug drawing features. Similar to the state sensor or
collision sensor, there is a list of actor tags `debug_tags` that needs
to be populated in the waypoint sensor. Anything in this list that has
waypoints will draw debug waypoints if the `draw_debug` flag is set.