Add CMake install to README
If CMake is installed with Visual Studio Code, the older version can cause issues when compiling. We need to explicitly install CMake to get everything working with VS Code and Visual Studio 2019. 

This was brought up as an issue when a customer had both Visual Studio 2017 and 2019 installed on the same machine. The version of CMake would continually pick up the wrong kit (2017) and when the kit was explicitly set it would say that VS 2019 was not supported.